+++ app/client/src/ce/workers/Evaluation/JSObject/JSProxy.ts
@@ -1,7 +1,7 @@
 import { isEmpty, set } from "lodash";
 import { MessageType, sendMessage } from "utils/MessageUtil";
-import { MAIN_THREAD_ACTION } from "../evalWorkerActions";
-import { isPromise } from "./utils";
+import { MAIN_THREAD_ACTION } from "@appsmith/workers/Evaluation/evalWorkerActions";
+import { isPromise } from "workers/Evaluation/JSObject/utils";
 
 export interface JSExecutionData {
   data: unknown;
@@ -13,13 +13,13 @@ export type JSFunctionProxy = (
   fullName: string,
 ) => unknown;
 
-export const JSFunctionProxyHandler = (
+export const jsFunctionProxyHandler = (
   fullName: string,
-  funcExecutionStart: () => void,
+  funcExecutionStart: (fullName: string) => void,
   funcExecutionEnd: (data: JSExecutionData) => void,
 ) => ({
   apply: function(target: any, thisArg: unknown, argumentsList: any) {
-    funcExecutionStart();
+    funcExecutionStart(fullName);
     let returnValue;
     try {
       returnValue = Reflect.apply(target, thisArg, argumentsList);
@@ -108,12 +108,12 @@ export class JSProxy {
 
   // Wrapper around JS Functions
   public JSFunctionProxy(
-    JSFunction: (...args: unknown[]) => unknown,
+    jsFunction: (...args: unknown[]) => unknown,
     jsFunctionFullName: string,
   ) {
     return new Proxy(
-      JSFunction,
-      JSFunctionProxyHandler(
+      jsFunction,
+      jsFunctionProxyHandler(
         jsFunctionFullName,
         this.functionExecutionStart,
         this.functionExecutionEnd,
