+++ app/client/src/ee/workers/Evaluation/JSObject/JSProxy.ts
@@ -1,0 +1,1 @@
+export * from "ce/workers/Evaluation/JSObject/JSProxy";
