+++ app/client/src/workers/Evaluation/PromisifyAction.ts
@@ -12,7 +12,7 @@ import _ from "lodash";
 import { EventType } from "constants/AppsmithActionConstants/ActionConstants";
 import { dataTreeEvaluator } from "./handlers/evalTree";
 import { TMessage, sendMessage, MessageType } from "utils/MessageUtil";
-import { MAIN_THREAD_ACTION } from "./evalWorkerActions";
+import { MAIN_THREAD_ACTION } from "@appsmith/workers/Evaluation/evalWorkerActions";
 import { ActionDescription } from "@appsmith/entities/DataTree/actionTriggers";
 
 export const promisifyAction = (
