+++ app/client/src/workers/Evaluation/handlers/__tests__/jsLibrary.test.ts
@@ -1,5 +1,5 @@
 import { installLibrary, uninstallLibrary } from "../jsLibrary";
-import { EVAL_WORKER_SYNC_ACTION } from "workers/Evaluation/evalWorkerActions";
+import { EVAL_WORKER_SYNC_ACTION } from "@appsmith/workers/Evaluation/evalWorkerActions";
 import * as mod from "../../../common/JSLibrary/ternDefinitionGenerator";
 
 jest.mock("../../../common/JSLibrary/ternDefinitionGenerator");
